+++ protocol/protocol.c
@@ -1,4 +1,38 @@
 #include "protocol.h"
+#include "buffer_builder.h"
+
+bool IsVarLenTask(uint8_t taskType)
+{
+  // FIXME: complete other conditions
+  if ( taskType == GET_OR_INSERT_REQ ) {
+    return true;
+  } else {
+    return false;
+  }
+}
+
+inline uint16_t GetTaskSize(void *task, bool isVarLen)
+{
+  // FIXME: complete other tasks
+  uint8_t taskType = ((Task*)task)->taskType;
+  uint16_t ret = sizeof(BlockDescriptorBase);
+
+  switch(taskType) {
+  case GET_OR_INSERT_REQ: {
+    GetOrInsertReq *req = (GetOrInsertReq*)task;
+    // |key + value + hash_id|
+    ret += req->len + sizeof(TupleIdT) + sizeof(uint32_t);
+    break;
+  }
+  default:
+    break;
+  }
+  if (isVarLen) {
+    ret += sizeof(uint32_t);
+  }
+  return ret;
+}
+
 
 void CreateCpuToDpuBufferForEachDPU()
 {
